refactor to use pinia store for conversation state
The chat widget endpoint is now using stores/chat.ts to manage the conversation state. The connection state is managed by stores/connection.ts.
OpeyController /chat proxy has been removed and the socket connection is managed by the store instead.
OpeyController /token proxy is used to get a jwt token from the server when the chat widget is loaded.
The token is used to authenticate the socket connection.